requirements.txt
openai==0.27.8
gradio==3.38.0
tiktoken==0.4.0
python-dotenv==1.0.1